feat: solve subtask modules
