+++ node-nodejs-basics/src/modules/index.js
@@ -1,0 +1,6 @@
+import { unknownObject } from './cjsToEsm.mjs';
+import { createMyServer } from './cjsToEsm.mjs';
+
+console.log(unknownObject);
+
+createMyServer.listen(8000);
